Rename provider file.
