+++ src/App.js
@@ -1,6 +1,6 @@
 import React from "react";
 
-import CDLProviders from "./Providers/CDLProviders";
+import ASPProviders from "./Providers/ASPProviders";
 import {Viewer} from "./Viewer/Viewer";
 
 import "bootstrap/dist/css/bootstrap.min.css";
@@ -13,8 +13,8 @@ import "./App.scss";
  */
 export function App () {
     return (
-        <CDLProviders>
+        <ASPProviders>
             <Viewer/>
-        </CDLProviders>
+        </ASPProviders>
     );
 }
